add support for subcommands
