+++ go.mod
@@ -1,3 +1,3 @@
-module github.com/dominikwinter/genpno
+module github.com/dominikwinter/pno
 
 go 1.21.6
